+++ src/Resources/contao/languages/de/tl_wertungsportal_stats.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Felder
+ */
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['datum']    = ['Datum', 'Tag der Zählung'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['funktion'] = ['Funktion', 'Interne Bezeichnung der Schnittstellenfunktion'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['endpunkt'] = ['Schnittstelle', 'Pfad der Schnittstellenfunktion'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['quelle']   = ['Quelle', 'Woher die Antwort kam'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['anzahl']   = ['Abrufe', 'Anzahl der Abrufe an diesem Tag'];
+
+/**
+ * Referenzen
+ */
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['quellen'] = [
+    'api'   => 'von der API',
+    'cache' => 'aus dem Cache',
+];
+
+/**
+ * Buttons
+ */
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['show'] = ['Details', 'Details des Datensatzes ID %s anzeigen'];