+++ src/Resources/contao/languages/en/tl_wertungsportal_stats.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Felder
+ */
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['datum']    = ['Date', 'Day of the count'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['funktion'] = ['Function', 'Internal name of the API function'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['endpunkt'] = ['Endpoint', 'Path of the API function'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['quelle']   = ['Source', 'Where the answer came from'];
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['anzahl']   = ['Requests', 'Number of requests on this day'];
+
+/**
+ * Referenzen
+ */
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['quellen'] = [
+    'api'   => 'from the API',
+    'cache' => 'from the cache',
+];
+
+/**
+ * Buttons
+ */
+$GLOBALS['TL_LANG']['tl_wertungsportal_stats']['show'] = ['Details', 'Show details of record ID %s'];